feat(receiver-mock): always parse X-Sumo-Fields in receiver
