+++ src/rust/receiver-mock/src/main.rs
@@ -15,6 +15,7 @@ mod logs;
 mod metrics;
 mod options;
 use options::Options;
+mod metadata;
 mod router;
 mod time;
